backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==3.2.25
django-cors-headers==4.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.8.0
idna==3.11
PyJWT==2.11.0
pymongo==3.12.3
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
sqlparse==0.2.4
urllib3==2.6.3
gunicorn==21.2.0
whitenoise==6.6.0
